Fixed motor control functions. Cleaned up formatting and type declarations.

+++ firmware/source/analogio.c
@@ -7,18 +7,17 @@
 #include "analogio.h"
 #include "util.h"
 
-
-
-void aio_init() {
+void aio_init()
+{
     adc_init();
     dma_init();
 }
 
-
-void adc_init() {
-    unsigned int config1, config2, config3, config4;
-    unsigned int configport_l, configport_h;
-    unsigned int configscan_l, configscan_h;
+void adc_init()
+{
+    uint16 config1, config2, config3, config4;
+    uint16 configport_l, configport_h;
+    uint16 configscan_l, configscan_h;
 
 
     config1 = ADC_MODULE_ON & ADC_IDLE_STOP & ADC_ADDMABM_ORDER & ADC_AD12B_12BIT &
@@ -38,17 +37,18 @@ void adc_init() {
     _TRISB2 = 1;
     _TRISB3 = 1;
     OpenADC1(config1, config2, config3, config4,
-            configport_l, configport_h,
-            configscan_h, configscan_l);
+             configport_l, configport_h,
+             configscan_h, configscan_l);
     SetChanADC1(0, ADC_CH0_NEG_SAMPLEB_VREFN & ADC_CH0_POS_SAMPLEB_AN1 &
-            ADC_CH0_NEG_SAMPLEA_NVREF & ADC_CH0_POS_SAMPLEA_AN1);
+                ADC_CH0_NEG_SAMPLEA_NVREF & ADC_CH0_POS_SAMPLEA_AN1);
 }
 
 
 int ADC_buf[ADC_NCHAN] __attribute__((space(dma)));
 
-void dma_init() {
-    unsigned int config, irq, sta_address, pad_address, count;
+void dma_init()
+{
+    uint16 config, irq, sta_address, pad_address, count;
 
     config = DMA0_MODULE_ON & DMA0_SIZE_WORD & PERIPHERAL_TO_DMA0 & DMA0_INTERRUPT_BLOCK &
             DMA0_NORMAL & DMA0_REGISTER_POST_INCREMENT & DMA0_CONTINUOUS;
@@ -62,9 +62,9 @@ void dma_init() {
     OpenDMA0(config, irq, sta_address, 0, pad_address, count);
 }
 
-
-float aio_read(unsigned int chan) {
-    return (float)ADC_buf[chan] / (float)ADC_RES;
+float aio_read(uint8 chan)
+{
+    return (float) ADC_buf[chan] / (float) ADC_RES;
 }
